Function finance::am

pub fn am(principal: f64, rate: f64, period: f64, year_or_month: bool, pay_at_beginning: bool) -> f64

In business, amortization refers to spreading payments over multiple periods. The term is used for two separate processes: amortization of loans and assets. It also refers to allocating the cost of an intangible asset over a period of time. see definition: https://en.wikipedia.org/wiki/Amortization_%28business%29