Function finance::roi

pub fn roi(cf0: f64, earnings: f64) -> f64

Return on Investment (ROI) is the benefit to an investor resulting from an investment of some resource. A high ROI means the investment gains compare favorably to investment cost. As a performance measure, ROI is used to evaluate the efficiency of an investment or to compare the efficiency of a number of different investments. In purely economic terms, it is one way of considering profits in relation to capital invested. see definition: https://en.wikipedia.org/wiki/Return_on_investment